feat: ErrorOr<TValue> is now a proper value object
This commit implements Equals and GetHashCode for the
ErrorOr<TValue> struct so that a deep comparison with the
errors list is performed (instead of the default reference
equality comparison). I introduced Microsoft.Bcl.HashCode for
.NET Standard 2.0 because I did not want to reinvent the wheel
for the HashCode class. If we do not want this dependency
then we should copy over the source code for .NET Standard 2.0.

Signed-off-by: Kenny Pflug <kenny.pflug@live.de>