feat: introduce KnownErrors
This commit introduces a new class called 'KnownErrors' which
contains the NoFirstError and NoErrors instances that
previously resided in ErrorOr<TValue>. With this change, they
will not reside in memory several times during runtime (once
for each type TValue is resolved to). Furthermore, I cached
two lists so that they won't be instantiated every time the
Errors and ErrorsOrEmptyList property is called.

Signed-off-by: Kenny Pflug <kenny.pflug@live.de>