perf: IsError is now a computed property
This change saves us 8 bits in the ErrorOr<TValue> struct which leads to less execution time when an instance is copied by value. As the struct is now properly encapsulated from a Design by Contract point of view, I simply choose _errors being null as the indicator for errors not being present.

Signed-off-by: Kenny Pflug <kenny.pflug@live.de>